ai_powerd_document/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
google-generativeai==0.8.3
PyMuPDF==1.24.11
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
python-dotenv==1.0.1
